Added the QRcode generation feature of the system. The post method generates the QRcode in the database. It then returns the QRcode image to the client. It also doesn't allow multiple qr codes for the same account. The get method resturns the QRcode image of the associated account. It also checks if the QRcode is already expired.
Added the logging feature to the system. The post method takes in a string from the request header which is the QRcode contents. It then creates an Attendance instance for the associated member in the QRcode. Another post request will then fill up the timeOut field of the Attendance instance. The get method returns the number of attendance instance with date the same as today and the timeOut is blank.

I have also changed the structure of the models and database to accommodate the new features.