Applied the REST framework to the Django system. Changed the codes for the attendance and account apps to fit in the REST framework. Added the url routes for the token acquisition and token refreshing.
Added the API for the Program and Workout modules. They utilize the IsAuthenticated permissions so the Access Token from the Json Web Token is required to get access to the API.

The program API provides access to the member's programs data. It allows the user to create, read, delete, and update the programs database.

The work API provides access to the member's workout data within the selected program. It allows the user to create, read, delete, and update the workouts database.

NOTE: Read the urls.py in the gym directory to know the url pattern for accessing the API.