Added the front-end for the email verfication and account registration pages. I have also added the api for account update for when the user wants to update their account details in the application.
I have also changed the format of the data returned by the program api. It now returns an array for programs with an array containing the workouts on that specific program.

I have also added the api for getting the program for the current day. It will return an empty JSON if the day does not have any program assigned to it

NOTE: CHECK URLS.PY FILES OF THE CORRESPONDING MODULES TO SEE THE URL PATTERN IN ACCESSING THEM.