I have created the api views for retrieving the trainer and member information through the usernames. It is needed when the api requires the user as a parameter.
I have also created the a Trainer model to the trainer's information. The member and schedule models references on the trainer model.

I have also created the API views for the scheduling system. The user can retrieve, edit, and delete the schedules with the trainer. Only the trainer can create the schedule with a member. Check the urls.py for information on access the API. TAKE NOTE: the API for listing out the schdules requires a query parameter together with the URL to properly retrieve the list of schedules. Example /api/scheduling/schedule?trainee=1 (the "1" here is the id of the member)
/api/scheduling/schedule?gymTrainer=0 if you want to access from the trainer side